fixing test for whe API fails
