+++ src/hooks/__tests__/useUserPermissions.test.tsx
@@ -1,27 +1,13 @@
-import React, { ReactNode } from 'react';
-import { QueryClient, QueryClientProvider } from 'react-query';
+import React from 'react';
 import useUserPermissions from '../useUserPermissions';
 import * as PlatformServices from '../../utilities/platformServices';
 import { renderHook } from '@testing-library/react-hooks';
 import fetch, { enableFetchMocks } from 'jest-fetch-mock';
+import { createWrapper } from '../../utilities/testHelpers';
 
 jest.mock('../../utilities/platformServices');
 
 enableFetchMocks();
-
-beforeEach(() => {
-  fetch.resetMocks();
-});
-
-const queryClient = new QueryClient();
-
-interface wrapperProps {
-  children: ReactNode;
-}
-
-const wrapper = ({ children }: wrapperProps) => (
-  <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
-);
 
 describe('useUserPermissions hook', () => {
   it('gets the user permissions back from two API calls', async () => {
@@ -37,10 +23,25 @@ describe('useUserPermissions hook', () => {
       }
     });
 
-    const { result, waitFor } = renderHook(() => useUserPermissions(), { wrapper });
+    const { result, waitFor } = renderHook(() => useUserPermissions(), {
+      wrapper: createWrapper()
+    });
 
     await waitFor(() => result.current.isSuccess);
 
     expect(result.current.data).toEqual({ isOrgAdmin: true, isSCACapable: true });
+  });
+
+  it('does not return anything if the API call fails', async () => {
+    const { authenticateUser } = PlatformServices;
+    (authenticateUser as jest.Mock).mockRejectedValue({ status: 'error' });
+
+    const { result, waitFor } = renderHook(() => useUserPermissions(), {
+      wrapper: createWrapper()
+    });
+
+    await waitFor(() => result.current.isSuccess);
+
+    expect(result.current.data).toEqual(undefined);
   });
 });
