+++ src/utilities/testHelpers.tsx
@@ -1,0 +1,16 @@
+import React, { ReactNode } from 'react';
+import { QueryClient, QueryClientProvider } from 'react-query';
+import { WrapperComponent } from '@testing-library/react-hooks';
+
+interface wrapperProps {
+  children: ReactNode;
+}
+
+export const createWrapper = (): WrapperComponent<unknown> => {
+  const queryClient = new QueryClient();
+  const wrapper = ({ children }: wrapperProps) => (
+    <QueryClientProvider client={queryClient}>{children}</QueryClientProvider>
+  );
+
+  return wrapper;
+};
